Fixed testing issue (#78)
The options testing was using `CommandLineArguments.Parse` after content
was deleted off of disk. That fails for certain operations like when a
ruleset file is used. To fix this just moved the option round trip
testing to the point of creation of the compiler log and that fixed the
issue.