Path normalization
When rehydrating on an OS with different paths from where the log was
created the rehydration will normalize the paths for the current OS.

This is necessary for the compilation to function correctly. Items like
parsing analyzer config options, classification, etc ... expect paths
for the current OS and will fail otherwise.