Add analyzer assembly info printing
It would be useful to be able to quickly dump detailed information about analyzer versions from complogs, so we can immediately see, for example, what git commit of the razor compiler was used. This adds just that; we look for the assembly name and AssemblyInformationalVersionAttribute for each analyzer and record that information in the log for printing. We also do this for all references, since it goes through the same code path, but no display of that information is currently done.