Fix bug around unix paths
Unix paths were accidentally being treated as options when exporting the
RSP file. This lead to the RSP file being invalid as it had paths that
weren't available the local machine (or even valid paths when crossing
between operating systems).

The fix was fairly simple but testing it caused a decent amount of churn
in the repository. Had to add APIs to make it possible to build up a
compiler log in memory instead of off of disk.