Supporting workspaces on giant binlogs (#172)
The old method of converting a binlog to a Workspace was by first converting to a compiler log. That process is fine for small / medium logs but for gigantic logs that can be quite expensive. Was attempting this on a binlog that had 1,700+ compilations inside it and the process quickly OOM'd.

This PR addresses the problem by loading Workspace directly from a binlog. This moves more methods into ICompilerCallReader and makes that the basis for SolutionReader. This was a pretty substantial refactoring of the code base as it required sharing a lot more concepts between the two readers.