Change in memory loading to better reflect how roslyn loads analyzers / generators (#209)
This change brings the in memory loader in line with how AnalyzerFileReference processes analyzers and generators. The code now uses System.Reflection.Metadata to scan for analyzer types in the assembly and then load only those into the process. This allows analyzer assemblies that have missing references that don't impact analyzers directly to still function. It also reduces the number of first chance exceptions generated in that scenario which significantly improves debugging.