Tolerance for missing files (#218)
* Tolerance for missing files

This changes compiler logs to be more tolerant of missing disk
artifacts. The old strategy was to simply error and bail out when
artifacts were missing on disk. This was designed primarily to make sure
customers weren't creating and submitting _stale_ binary logs.

Diagnostics around potentially _stale_ logs are valuable but it was too
agressive. It meant that the tooling couldn't represent `Compilations`
with those types of errors. For example: instead of creating a
`Compilation` where emit fails because of a missing key file the tooling
just could not create a `Compilation` at all.

This change is pushing the tooling to be tolerant of the missing
artifacts while preserving the _stale_ diagnostics in `complog create`.

closes #217