Move arguments storage into the readers (#288)
* Moved bulk of GetArguments calls

Eventual goal is for this to be implemented in a normalized fashion inside of `CompilerLogReader`. This moves the bulk of the calls to going through those types. First step in the refactoring.

* Name refctaro

* Refactor binary log reading

Now a tuple of `CompilerCall` and the underlying arguments string.

* Quick rename

* Pushed args up one more layer

* Removed bad extension methods

* Removed GetArguments

Big milestone in that the arguments are no longer stored on compiler call. That means that the implemantion of `CompilerCallReader` now owns how it is represented and can do some normalization.

* more

* more

* Remove end_of_line setting from .editorconfig (#290)

* fix unix issue

* more

* more

* more